+++ stdlib/lib.core.d.ts
@@ -1845,6 +1845,10 @@ declare module "node:events" {
 type DnsLookupCallback = (err: any, address: string, family: number) => void;
 type DnsLookupAllCallback = (err: any, addresses: any[]) => void;
 type DnsLookupFamily = 0 | 4 | 6;
+type DnsResolveCallback = (err: any, addresses: string[]) => void;
+interface DnsResolveOptions {
+    ttl?: boolean;
+}
 interface DnsLookupOptions {
     family?: DnsLookupFamily;
     all?: boolean;
@@ -1855,6 +1859,8 @@ interface DnsLookupOptions {
 interface DnsPromises {
     lookup(hostname: string): Promise<any>;
     lookup(hostname: string, options: DnsLookupOptions | DnsLookupFamily | undefined, ...ignored: any[]): Promise<any>;
+    resolve4(hostname: string): Promise<string[]>;
+    resolve4(hostname: string, options: DnsResolveOptions | undefined, ...ignored: any[]): Promise<string[]>;
 }
 interface DNS {
     readonly ADDRCONFIG: number;
@@ -1864,6 +1870,8 @@ interface DNS {
     lookup(hostname: string, callback: DnsLookupCallback, ...ignored: any[]): void;
     lookup(hostname: string, options: DnsLookupOptions | DnsLookupFamily | undefined, callback: DnsLookupCallback, ...ignored: any[]): void;
     lookup(hostname: string, options: DnsLookupOptions | DnsLookupFamily | undefined, callback: DnsLookupAllCallback, ...ignored: any[]): void;
+    resolve4(hostname: string, callback: DnsResolveCallback, ...ignored: any[]): void;
+    resolve4(hostname: string, options: DnsResolveOptions | undefined, callback: DnsResolveCallback, ...ignored: any[]): void;
 }
 declare const dns: DNS;
 declare module "dns" {
@@ -1874,6 +1882,8 @@ declare module "dns" {
     export function lookup(hostname: string, callback: DnsLookupCallback, ...ignored: any[]): void;
     export function lookup(hostname: string, options: DnsLookupOptions | DnsLookupFamily | undefined, callback: DnsLookupCallback, ...ignored: any[]): void;
     export function lookup(hostname: string, options: DnsLookupOptions | DnsLookupFamily | undefined, callback: DnsLookupAllCallback, ...ignored: any[]): void;
+    export function resolve4(hostname: string, callback: DnsResolveCallback, ...ignored: any[]): void;
+    export function resolve4(hostname: string, options: DnsResolveOptions | undefined, callback: DnsResolveCallback, ...ignored: any[]): void;
     const defaultDns: DNS;
     export default defaultDns;
 }
@@ -1885,16 +1895,20 @@ declare module "node:dns" {
     export function lookup(hostname: string, callback: DnsLookupCallback, ...ignored: any[]): void;
     export function lookup(hostname: string, options: DnsLookupOptions | DnsLookupFamily | undefined, callback: DnsLookupCallback, ...ignored: any[]): void;
     export function lookup(hostname: string, options: DnsLookupOptions | DnsLookupFamily | undefined, callback: DnsLookupAllCallback, ...ignored: any[]): void;
+    export function resolve4(hostname: string, callback: DnsResolveCallback, ...ignored: any[]): void;
+    export function resolve4(hostname: string, options: DnsResolveOptions | undefined, callback: DnsResolveCallback, ...ignored: any[]): void;
     const defaultDns: DNS;
     export default defaultDns;
 }
 declare module "dns/promises" {
     export const lookup: DnsPromises["lookup"];
+    export const resolve4: DnsPromises["resolve4"];
     const defaultDnsPromises: DnsPromises;
     export default defaultDnsPromises;
 }
 declare module "node:dns/promises" {
     export const lookup: DnsPromises["lookup"];
+    export const resolve4: DnsPromises["resolve4"];
     const defaultDnsPromises: DnsPromises;
     export default defaultDnsPromises;
 }